old/requirements.txt
discord.py
async-timeout
requests
psycopg2-binary
sqlalchemy
apscheduler